feat(ClientFactory): Allow for API specific client classes
