+++ tests/auth_test.py
@@ -4,21 +4,10 @@
 import datetime
 from unittest import mock
 
-import flask
 import pytest
 import requests_mock as rm
 
 from impact_stack.rest import AuthMiddleware, ClientFactory
-
-
-@pytest.fixture(name="app", scope="class")
-def fixture_app():
-    """Define a test Flask app."""
-    app = flask.Flask("test-app")
-    app.config["IMPACT_STACK_API_URL"] = "https://impact-stack.net/api"
-    app.config["IMPACT_STACK_API_KEY"] = "api-key"
-    with app.app_context():
-        yield app
 
 
 def test_auth_client(app, requests_mock):
