+++ tests/conftest.py
@@ -1,0 +1,14 @@
+"""Define fixtures."""
+
+import flask
+import pytest
+
+
+@pytest.fixture(name="app", scope="class")
+def fixture_app():
+    """Define a test Flask app."""
+    app = flask.Flask("test-app")
+    app.config["IMPACT_STACK_API_URL"] = "https://impact-stack.net/api"
+    app.config["IMPACT_STACK_API_KEY"] = "api-key"
+    with app.app_context():
+        yield app
